+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 - Add option to install with homebrew
 - Added `condition` option to tasks
+- Added `exclude` tera filter
+- Added `input` tera function
 
 ## v1.0.0 - 2023-05-23